k8s: remove unused converter functions
These are unused since commit c2b935042f3e ("bgp: remove metallb bgp
integration").
