+++ pkg/k8s/factory_functions_test.go
@@ -10,7 +10,6 @@ import (
 
 	"github.com/stretchr/testify/require"
 	core_v1 "k8s.io/api/core/v1"
-	networkingv1 "k8s.io/api/networking/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/tools/cache"
 
@@ -867,163 +866,6 @@ func Test_EqualV1Namespace(t *testing.T) {
 	for _, tt := range tests {
 		got := tt.args.o1.DeepEqual(tt.args.o2)
 		require.Equal(t, tt.want, got, "Test Name: %s", tt.name)
-	}
-}
-
-func Test_ConvertToK8sV1ServicePorts(t *testing.T) {
-	type args struct {
-		ports []slim_corev1.ServicePort
-	}
-	tests := []struct {
-		name string
-		args args
-		want []core_v1.ServicePort
-	}{
-		{
-			name: "empty",
-			args: args{
-				ports: []slim_corev1.ServicePort{},
-			},
-			want: []core_v1.ServicePort{},
-		},
-		{
-			name: "non-empty",
-			args: args{
-				ports: []slim_corev1.ServicePort{
-					{
-						Name: "foo",
-						Port: int32(1),
-					},
-				},
-			},
-			want: []core_v1.ServicePort{
-				{
-					Name: "foo",
-					Port: int32(1),
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		got := ConvertToK8sV1ServicePorts(tt.args.ports)
-		require.Equalf(t, tt.want, got, "Test Name: %s", tt.name)
-	}
-}
-
-func Test_ConvertToK8sV1SessionAffinityConfig(t *testing.T) {
-	ts := int32(1)
-	type args struct {
-		cfg *slim_corev1.SessionAffinityConfig
-	}
-	tests := []struct {
-		name string
-		args args
-		want *core_v1.SessionAffinityConfig
-	}{
-		{
-			name: "empty",
-			args: args{
-				cfg: &slim_corev1.SessionAffinityConfig{},
-			},
-			want: &core_v1.SessionAffinityConfig{},
-		},
-		{
-			name: "non-empty",
-			args: args{
-				cfg: &slim_corev1.SessionAffinityConfig{
-					ClientIP: &slim_corev1.ClientIPConfig{
-						TimeoutSeconds: &ts,
-					},
-				},
-			},
-			want: &core_v1.SessionAffinityConfig{
-				ClientIP: &core_v1.ClientIPConfig{
-					TimeoutSeconds: &ts,
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		got := ConvertToK8sV1ServiceAffinityConfig(tt.args.cfg)
-		require.Equalf(t, tt.want, got, "Test Name: %s", tt.name)
-	}
-}
-
-func Test_ConvertToK8sV1LoadBalancerIngress(t *testing.T) {
-	type args struct {
-		ings []slim_corev1.LoadBalancerIngress
-	}
-	tests := []struct {
-		name string
-		args args
-		want []core_v1.LoadBalancerIngress
-	}{
-		{
-			name: "empty",
-			args: args{
-				ings: []slim_corev1.LoadBalancerIngress{},
-			},
-			want: []core_v1.LoadBalancerIngress{},
-		},
-		{
-			name: "non-empty",
-			args: args{
-				ings: []slim_corev1.LoadBalancerIngress{
-					{
-						IP: "1.1.1.1",
-					},
-				},
-			},
-			want: []core_v1.LoadBalancerIngress{
-				{
-					IP:    "1.1.1.1",
-					Ports: nil,
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		got := ConvertToK8sV1LoadBalancerIngress(tt.args.ings)
-		require.Equalf(t, tt.want, got, "Test Name: %s", tt.name)
-	}
-}
-
-func Test_ConvertToNetworkV1IngressLoadBalancerIngress(t *testing.T) {
-	type args struct {
-		ings []slim_corev1.LoadBalancerIngress
-	}
-	tests := []struct {
-		name string
-		args args
-		want []networkingv1.IngressLoadBalancerIngress
-	}{
-		{
-			name: "empty",
-			args: args{
-				ings: []slim_corev1.LoadBalancerIngress{},
-			},
-			want: []networkingv1.IngressLoadBalancerIngress{},
-		},
-		{
-			name: "non-empty",
-			args: args{
-				ings: []slim_corev1.LoadBalancerIngress{
-					{
-						IP: "1.1.1.1",
-					},
-				},
-			},
-			want: []networkingv1.IngressLoadBalancerIngress{
-				{
-					IP:    "1.1.1.1",
-					Ports: []networkingv1.IngressPortStatus{},
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		got := ConvertToNetworkV1IngressLoadBalancerIngress(tt.args.ings)
-		require.Equalf(t, tt.want, got, "Test Name: %s", tt.name)
 	}
 }
